+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - change `sj_per_unit(j::String, net::Network{MultiPhase})` to return complex values
 - change `ConstraintInfo.set_type::MOI.AbstractSet` -> `ConstraintInfo.set_type::DataType`
 - add `Units.RadiansUnit` and `Dimensions.BusTerminalDimension`
-- add psse .raw to Network parsing
+- start psse .raw to Network parsing
 - add `Network_IEEE118`
 
 ## v0.4.7